htmlrenderer implementation correction
